plsql_Acc_backend/requirements.txt
# Core Python dependencies
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.0

# PL/SQL Parsing and AST Generation
antlr4-python3-runtime>=4.13.0
sqlparse>=0.4.4

# Dependency Analysis and Graph Generation
networkx>=3.1
pygraphviz>=1.11

# LLM Integration
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.0.350
langchain-community>=0.0.25

# Template Engine
jinja2>=3.1.2

# Database Connectivity
oracledb
cx-Oracle==8.3.0
setuptools==80.10.2
pymysql>=1.1.0
psycopg2-binary>=2.9.7

# Testing and Validation
pytest>=7.4.0
pytest-asyncio>=0.21.0
testcontainers>=3.7.0

# Configuration and Utilities
pyyaml>=6.0.1
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Logging and Monitoring
structlog>=23.1.0
rich>=13.5.0

# Git Integration
GitPython>=3.1.36

# Performance and Caching
redis>=4.6.0
asyncio-throttle>=1.0.2

# CLI and Progress
tqdm>=4.65.0
typer>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# File Processing
pathlib2>=2.3.7
watchdog>=3.0.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Enterprise Features
kafka-python>=2.0.2
pika>=1.3.2

# Development Dependencies (optional)
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
